add: clean architecture module setup
