+++ domain/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    id("java-library")
+    id("org.jetbrains.kotlin.jvm")
+}
+
+java {
+    sourceCompatibility = Apps.javaCompileOption
+    targetCompatibility = Apps.javaCompileOption
+}
+
+dependencies {
+    implementation(project(":entity"))
+}
