+++ domain/src/main/java/com/marassu/domain/repository/UserRepository.kt
@@ -1,0 +1,10 @@
+package com.marassu.domain.repository
+
+import com.marassu.entity.user.UserLoginRequest
+import com.marassu.entity.user.UserRegisterRequest
+
+interface UserRepository {
+    suspend fun postUserRegister(userRegisterRequest: UserRegisterRequest)
+
+    suspend fun postUserLogin(userLoginRequest: UserLoginRequest)
+}
