+++ entity/build.gradle.kts
