+++ entity/src/main/java/com/marassu/entity/petition/PetitionRequest.kt
@@ -1,4 +1,4 @@
-package petition
+package com.marassu.entity.petition
 
 import com.google.gson.annotations.SerializedName
