+++ entity/src/main/java/com/marassu/entity/user/UserLoginRequest.kt
@@ -1,14 +1,11 @@
-package user
+package com.marassu.entity.user
 
 import com.google.gson.annotations.SerializedName
 
-data class UserRegisterRequest(
+data class UserLoginRequest(
     @SerializedName("email")
     val email: String,
 
     @SerializedName("password")
-    val password: String,
-
-    @SerializedName("name")
-    val name: String
+    val password: String
 )
