+++ entity/src/main/java/com/marassu/entity/user/UserRegisterRequest.kt
@@ -1,0 +1,14 @@
+package com.marassu.entity.user
+
+import com.google.gson.annotations.SerializedName
+
+data class UserRegisterRequest(
+    @SerializedName("email")
+    val email: String,
+
+    @SerializedName("password")
+    val password: String,
+
+    @SerializedName("name")
+    val name: String
+)
